+++ pubsub.go
@@ -5,6 +5,7 @@ import (
 	"sync"
 
 	events "github.com/kataras/go-events"
+	"github.com/scylladb/go-set/strset"
 )
 
 //PubSubManager 广播器
@@ -28,8 +29,8 @@ func NewPubSubManager() *PubSubManager {
 //ChannelInUse 检测channeldid是否在被使用
 func (psm *PubSubManager) ChannelInUse(channelid string) bool {
 	psm.lock.RLock()
+	defer psm.lock.RUnlock()
 	_, ok := psm.channels[channelid]
-	psm.lock.RUnlock()
 	return ok
 }
 
@@ -37,8 +38,8 @@ func (psm *PubSubManager) ChannelInUse(channelid string) bool {
 func (psm *PubSubManager) AddChannel(channelid string) {
 	if !psm.ChannelInUse(channelid) {
 		psm.lock.Lock()
+		defer psm.lock.Unlock()
 		psm.channels[channelid] = make(chan struct{})
-		psm.lock.Unlock()
 	}
 }
 
@@ -47,8 +48,8 @@ func (psm *PubSubManager) CloseChannel(channelid string) {
 	if psm.ChannelInUse(channelid) {
 		close(psm.channels[channelid])
 		psm.lock.Lock()
+		defer psm.lock.Unlock()
 		delete(psm.channels, channelid)
-		psm.lock.Unlock()
 	}
 	psm.emmiter.RemoveAllListeners(events.EventName(channelid))
 }
@@ -57,18 +58,20 @@ func (psm *PubSubManager) CloseChannel(channelid string) {
 func (psm *PubSubManager) Channels() []string {
 	result := []string{}
 	psm.lock.RLock()
-	for k, _ := range psm.channels {
+	defer psm.lock.RUnlock()
+	for k := range psm.channels {
 		result = append(result, k)
 	}
-	psm.lock.RUnlock()
+
 	return result
 }
 
-//Publish 指定频道广播消息
-//@Param msg interface{} 广播的消息
-//@Param channels ...string 广播的频道
-func (psm *PubSubManager) Publish(msg interface{}, channels ...string) {
-	for _, channelid := range channels {
+//Send 指定频道发送消息
+//@params msg interface{} 发送的消息
+//@params channels ...string 发送去的频道
+func (psm *PubSubManager) Send(msg interface{}, channels ...string) {
+	chaset := strset.New(channels...)
+	for _, channelid := range chaset.List() {
 		if !psm.ChannelInUse(channelid) {
 			psm.AddChannel(channelid)
 		}
@@ -77,23 +80,50 @@ func (psm *PubSubManager) Publish(msg interface{}, channels ...string) {
 	}
 }
 
-//PublishWithDefault 指定频道广播消息,消息一并广播到default这个channel上
-//@Param msg interface{} 广播的消息
-//@Param channels ...string 广播的频道
-func (psm *PubSubManager) PublishWithDefault(msg interface{}, channels ...string) {
+//SendWithDefault 指定频道发送消息,消息一并发送到default这个channel上
+//@params msg interface{} 发送的消息
+//@params channels ...string 发送去的频道
+func (psm *PubSubManager) SendWithDefault(msg interface{}, channels ...string) {
 	channels = append(channels, "default")
-	psm.Publish(msg, channels...)
+	psm.Send(msg, channels...)
+
+}
+
+//Publish 全频道广播消息
+//@params msg interface{} 广播的消息
+func (psm *PubSubManager) Publish(msg interface{}) {
+	channels := []string{}
+	psm.lock.RLock()
+	defer psm.lock.RUnlock()
+	for k := range psm.channels {
+		channels = append(channels, k)
+	}
+	psm.Send(msg, channels...)
+}
+
+//PublishWithoutDefault 除了default这个channel外全频道广播
+//@params msg interface{} 广播的消息
+func (psm *PubSubManager) PublishWithoutDefault(msg interface{}) {
+	channels := []string{}
+	psm.lock.RLock()
+	defer psm.lock.RUnlock()
+	for k := range psm.channels {
+		if k != "default" {
+			channels = append(channels, k)
+		}
+	}
+	psm.Send(msg, channels...)
 }
 
 type CloseListenerFunc func()
 
 //RegistListener 注册频道的监听器
-//@Param channelid string 频道id
-//@Param size int 构造的消息channel的缓冲区大小,<=0时使用无缓冲channel
-//@Returns <-chan interface{} 消息channel
-//@Returns <-chan struct{} 关闭监听器的信号channel
-//@Returns CloseListenerFunc 关闭当前注册监听器的函数
-//@Returns error 当频道未被注册可以使用时会报错
+//@params channelid string 频道id
+//@params size int 构造的消息channel的缓冲区大小,<=0时使用无缓冲channel
+//@returns <-chan interface{} 消息channel
+//@returns <-chan struct{} 关闭监听器的信号channel
+//@returns CloseListenerFunc 关闭当前注册监听器的函数
+//@returns error 当频道未被注册可以使用时会报错
 func (psm *PubSubManager) RegistListener(channelid string, size int) (<-chan interface{}, <-chan struct{}, CloseListenerFunc, error) {
 	if !psm.ChannelInUse(channelid) {
 		return nil, nil, nil, errors.New("channel not in use now")
@@ -119,13 +149,13 @@ func (psm *PubSubManager) RegistListener(channelid string, size int) (<-chan int
 }
 
 //CloseNotify 关闭指定频道的信号提醒
-//@Param channelid string 频道id
-//@Returns <-chan struct{} 关闭频道所有监听器的信号channel
-//@Returns bool 当前频道是否可用
+//@params channelid string 频道id
+//@returns <-chan struct{} 关闭频道所有监听器的信号channel
+//@returns bool 当前频道是否可用
 func (psm *PubSubManager) CloseNotify(channelid string) (ch <-chan struct{}, ok bool) {
 	psm.lock.RLock()
+	defer psm.lock.RUnlock()
 	ch, ok = psm.channels[channelid]
-	psm.lock.RUnlock()
 	return
 }